extract printing stuff
