+++ packages/ai/src/evals/eval.types.ts
@@ -236,4 +236,6 @@ export type EvaluationReport = {
   };
 };
 
+// TODO: BEFORE MERGE - why are these two separate things? Very confusing....
 export type MetaWithEval = TaskMeta & { evaluation: EvaluationReport };
+export type MetaWithCase = TaskMeta & { case: EvalCaseReport };
